+++ pages/hit-and-blow/solver.tsx
@@ -1,45 +1,26 @@
 import React, { useState, useEffect } from "react";
 import type { NextPage } from "next";
 import Head from "next/head";
-import Switch from "@mui/material/Switch";
 import Grid from "@mui/material/Grid";
-import TextField from "@mui/material/TextField";
-import Button from "@mui/material/Button";
-import Table from "@mui/material/Table";
-import TableContainer from "@mui/material/TableContainer";
-import TableCell from "@mui/material/TableCell";
-import TableRow from "@mui/material/TableRow";
-import TableBody from "@mui/material/TableBody";
-import Paper from '@mui/material/Paper';
-import ToggleButton from '@mui/material/ToggleButton';
-import ToggleButtonGroup from '@mui/material/ToggleButtonGroup';
-import FormGroup from '@mui/material/FormGroup';
-import FormControlLabel from '@mui/material/FormControlLabel';
 
 import NavigationBar from "@/components/NavigationBar";
 import ContentsCard from "@/components/ContentsCard";
 import CenterrizedHorizontalGrid from "@/components/CenterrizedHorizontalGrid";
+import HitAndBlowDigitsSelector from "@/components/HitAndBlowDigitsSelector";
+import HitAndBlowAskTable from "@/components/HitAndBlowAskTable";
+import HitAndBlowForm from "@/components/HitAndBlowForm";
+import HitAndBlowResult from "@/components/HitAndBlowResult";
 
-import { HitCounter, BlowCounter, Answer } from "@/lib/hitandblow/general";
+import { HitCounter, BlowCounter, InitializeAnswer } from "@/lib/hitandblow/general";
+import { History } from "@/types/HitAndBlow";
 
 const Solver: NextPage = () => {
-
-  interface History {
-    ask: number[];
-    hit: number;
-    blow: number;
-  }
-
   const [numberlength, setNumberlength] = useState<number>(3);
   const [history, setHistory] = useState<History[]>([]);
-  const [askednumber, setAskednumber] = useState<string>("");
-  const [hit, setHit] = useState<string>("");
-  const [blow, setBlow] = useState<string>("");
-  const [candidate, setCandidate] = useState<number[][]>(Answer(numberlength))
-  const [displayall, setDisplayall] = useState<boolean>(false);
+  const [candidate, setCandidate] = useState<number[][]>(InitializeAnswer(numberlength))
 
   useEffect(() => {
-    setCandidate(Answer(numberlength));
+    setCandidate(InitializeAnswer(numberlength));
   }, [numberlength]);
 
   const handleNumberLength = (
@@ -48,31 +29,12 @@ const Solver: NextPage = () => {
   ) => {
     setNumberlength(newNumberLength);
   };
-  const toggleSwitch = () => {
-    setDisplayall(!displayall)
-  }
 
-  const handleApply = () => {
-    const asknum = askednumber.split("").map((e) => parseInt(e));
-    const hitnum = parseInt(hit) || 0;
-    const blownum = parseInt(blow) || 0;
-
-    if (hitnum + blownum > numberlength) return;
-    if (askednumber.length !== numberlength) return;
-
-    const hist: History = {
-      ask: asknum,
-      hit: hitnum,
-      blow: blownum,
-    }
+  const addHistory = (newHistory: History) => {
     setCandidate(candidate
-      .filter((ans) => HitCounter(ans, asknum) === hitnum)
-      .filter((ans) => BlowCounter(ans, asknum) === blownum));
-
-    setHistory([...history, hist]);
-    setAskednumber("")
-    setHit("")
-    setBlow("")
+      .filter((cand) => HitCounter(cand, newHistory.ask) === newHistory.hit)
+      .filter((cand) => BlowCounter(cand, newHistory.ask) === newHistory.blow));
+    setHistory([...history, newHistory]);
   }
 
   return (
@@ -87,61 +49,10 @@ const Solver: NextPage = () => {
         <Grid container spacing={{ xs: 2, md: 3 }} columns={12}>
           <Grid item xs={12}>
             <ContentsCard>
-              <ToggleButtonGroup
-                value={numberlength}
-                exclusive
-                onChange={handleNumberLength}
-              >
-                <ToggleButton value={3}>
-                  3
-                </ToggleButton>
-                <ToggleButton value={4}>
-                  4
-                </ToggleButton>
-                <ToggleButton value={5}>
-                  5
-                </ToggleButton>
-              </ToggleButtonGroup>
-              <TableContainer component={Paper}>
-                <Table>
-                  <TableBody>
-                    {history.map((h) => (
-                      <TableRow key={h.ask.join('')}>
-                        <TableCell>
-                          {h.ask.join('')}
-                        </TableCell>
-                        <TableCell>
-                          {h.hit} hit
-                        </TableCell>
-                        <TableCell>
-                          {h.blow} blow
-                        </TableCell>
-                      </TableRow>
-                    ))}
-                  </TableBody>
-                </Table>
-              </TableContainer>
-              <div>
-                <TextField label="number" value={askednumber} onChange={e => setAskednumber(e.target.value)} />
-                <TextField label="hit" value={hit} onChange={e => setHit(e.target.value)} />
-                <TextField label="blow" value={blow} onChange={e => setBlow(e.target.value)} />
-                <Button variant="contained" onClick={handleApply}>Apply</Button>
-              </div>
-              <Grid container spacing={{ xs: 2 }} columns={12}>
-                {candidate.map((value, i) => {
-                  if (!displayall && i > 1000) return;
-                  return (
-                    <Grid item key={value.join('')}>
-                      {value}
-                    </Grid>
-                  );
-                })
-                }
-              </Grid>
-              <FormGroup>
-                <FormControlLabel control={
-                  <Switch checked={displayall} onChange={toggleSwitch} />} label="display all" />
-              </FormGroup>
+              <HitAndBlowDigitsSelector numberlength={numberlength} handleNumberLength={handleNumberLength} />
+              <HitAndBlowAskTable history={history} />
+              <HitAndBlowForm numberlength={numberlength} addHistory={addHistory} />
+              <HitAndBlowResult candidate={candidate} />
             </ContentsCard>
           </Grid>
         </Grid>
