+++ types/HitAndBlow.d.ts
@@ -1,0 +1,5 @@
+export interface History {
+  ask: number[];
+  hit: number;
+  blow: number;
+}
